try build and run approach
